Can take picture of book and upload to Db. GeoLocation class cleaned up.
From the MyBook activity, you can now take a picture of a book and upload it directly to firebase (for testing). No DB integration yet. GeoLocation code cleaned up.  If permissions are denied for either camera or location access, a message appears saying so. You can access GeoLocation from the "SELECT ADDRESS FROM MAP" button in "Profile". To take a picture of a book (for now), click "My Books", "TemporaryBook".