Removed views in myBook. Can now get address from lat/lng.
Views in the myBook activity that pertain to taking a picture of a book and saving it are temporarily commented out. Selecting an address from the map for your profile now saves the actual address, not just the lat/lng. Marker now displays where pickup location is for a book.