Added feature to take a picture and view a picture of a book
When you create a new book or when you edit your own book, you are able to take a picture of it and link it to the book. When you edit your own book or view someone else's book (assuming both have a picture), you are able to view the picture that the owner took. Removing a book removes the book's picture from firebase. The "Image image" variable from the book class, as well as the appropriate getter and setter have been removed, as they are no longer necessary.